Fixed bug causing worlds to not be connected
When the game was retrofitted to only allow connections to the original
world, and the number of pastes returned was more than 4, the downloader
thread would lock the world queue and enter an infinite loop attempting to
generate a position for the new world. Additionally, due to the reduced
sleep time I was using for debugging, this deadlock would occur before any
worlds were connected. Fixed by returning from downloader thread when
world limit is exceeded.