Fixed some minor bugs with world select
If you deleted a world or added a new one, entered a world, and went back
to the menu, the options would be collated and alphabetized but the
buttons would still display their old text. This could cause a crash.
Fixed. (Although when you leave and re-enter your saves are still
alphabetized and collated)