Fixed character not jumping correctly
When jumping along the positive Y-axis or along the axis defined by
(-0.5, 0.5) the jump would be about half as powerful as it's supposed to
be, as well as exhibiting some other odd behavior. This was caused by a
line setting the character's view direction to the camera's world matrix's
forward; which has approximately the same value as the character's Down
orientation. Switching this to the View matrix's forward worked fine.